# apiGateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway для микросервисного приложения вишлистов",
        "title": "Wishlist API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8095",
    "basePath": "/api/v1",
    "paths": {
...
        "/items/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает предмет по его ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Получить предмет",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID предмета",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Предмет найден",
                        "schema": {
                            "$ref": "#/definitions/handlers.ItemDTO"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет существующий предмет (только владелец вишлиста)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Обновить предмет",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID предмета",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateItemRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Предмет обновлен",
                        "schema": {
                            "$ref": "#/definitions/handlers.ItemDTO"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет предмет из вишлиста (только владелец)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Удалить предмет",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID предмета",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешное удаление"
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/items/{id}/book": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Пользователь бронирует предмет в чужом вишлисте",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Забронировать предмет",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID предмета",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Предмет забронирован",
                        "schema": {
                            "$ref": "#/definitions/handlers.ItemDTO"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/items/{id}/unbook": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Отменяет бронирование предмета (бронировавший или владелец)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Отменить бронирование",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID предмета",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бронирование отменено",
                        "schema": {
                            "$ref": "#/definitions/handlers.ItemDTO"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error содержит строку с расшифровкой того что пошло не так\nminLength: 1\nmaxLength: 100\nexample: item not found",
                    "type": "string"
                }
            }
        },
        "handlers.ItemDTO": {
            "type": "object",
            "properties": {
                "booked_at": {
                    "description": "BookedAt Дата и время бронирования, если есть\nformat: date-time\nnullable: true\nexample: 1700000000",
                    "type": "integer"
                },
                "booked_by": {
                    "description": "BookedBy UUID пользователя, забронировавшего предмет, если забронирован\nformat: uuid\nnullable: true\nexample: 123e4567-e89b-12d3-a456-426614174000",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt Дата и время создания (Unix timestamp)\nformat: date-time\nexample: 1700000000",
                    "type": "integer"
                },
                "description": {
                    "description": "Description Описание айтема\nmaxLength: 1000\nexample: Сертификат, желательно на четверых, и чтобы красивый был и с топорами",
                    "type": "string"
                },
                "id": {
                    "description": "ID UUID айтема\nformat: uuid\nexample: 123e4567-e89b-12d3-a456-426614174000",
                    "type": "string"
                },
                "image_url": {
                    "description": "ImageURL URL изображения\nformat: uri\nexample: https://example.com/image.jpg",
                    "type": "string"
                },
                "name": {
                    "description": "Name Название айтема\nrequired: true\nminLength: 1\nmaxLength: 80\nexample: Сертификат на метание топоров",
                    "type": "string"
                },
                "price": {
                    "description": "Price Цена в копейках (или предполагаемая цена)\nminimum: 0\nexample: 50000",
                    "type": "integer"
                },
                "product_url": {
                    "description": "ProductURL URL на то где можно приобрести\nformat: uri\nexample: https://example.com/product/123",
                    "type": "string"
                },
                "updated_at": {
                    "description": "UpdatedAt Дата и время последнего обновления (Unix timestamp)\nformat: date-time\nexample: 1700000000",
                    "type": "integer"
                },
                "wishlist_id": {
                    "description": "ID UUID вишлиста в котором находится айтем\nformat: uuid\nexample: 123e4567-e89b-12d3-a456-426614174000",
                    "type": "string"
                }
            }
        },
...
        "handlers.UpdateItemRequestDTO": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "description": "Description Описание айтема\nmaxLength: 1000\nexample: Сертификат, желательно на четверых, и чтобы красивый был и с топорами",
                    "type": "string"
                },
                "image_url": {
                    "description": "ImageURL URL изображения\nformat: uri\nexample: https://example.com/image.jpg",
                    "type": "string"
                },
                "name": {
                    "description": "Name Название айтема\nrequired: true\nminLength: 1\nmaxLength: 80\nexample: Сертификат на метание топоров",
                    "type": "string"
                },
                "price": {
                    "description": "Price Цена в копейках (или предполагаемая цена)\nminimum: 0\nexample: 50000",
                    "type": "integer"
                },
                "product_url": {
                    "description": "ProductURL URL на то где можно приобрести\nformat: uri\nexample: https://example.com/product/123",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}